ref(tests): Refactor jest setup file (#9982)
This moves all tests data into modules in `tests/js/fixtures` and reorganizes the `setup.js` file a bit.

To add to window.TestStubs, create a module that exports a function (note, use named exports, NOT default!) in tests/js/fixtures and it will automatically be added. subdirectories are currently not supported.
