+++ tests/js/fixtures/accountAppearance.js
@@ -1,0 +1,9 @@
+export function AccountAppearance(params = {}) {
+  return {
+    stacktrace_order: '2',
+    timezone: 'US/Pacific',
+    language: 'en',
+    clock_24_hours: true,
+    ...params,
+  };
+}
