+++ tests/js/fixtures/authProviders.js
@@ -1,0 +1,3 @@
+export function AuthProviders(params = []) {
+  return [['dummy', 'Dummy', false], ...params];
+}
