+++ tests/js/fixtures/authProvidersSaml.js
@@ -1,0 +1,3 @@
+export function AuthProvidersSaml(params = []) {
+  return [['dummy', 'Dummy', true], ...params];
+}
