+++ tests/js/fixtures/eventIdQueryResult.js
@@ -1,0 +1,18 @@
+import {Event} from './event';
+
+export function EventIdQueryResult(params = {}) {
+  let event = Event({
+    metadata: {
+      type: 'event type',
+      value: 'event description',
+    },
+  });
+  return {
+    organizationSlug: 'org-slug',
+    projectSlug: 'project-slug',
+    groupId: event.groupID,
+    eventId: event.eventID,
+    event,
+    ...params,
+  };
+}
