+++ tests/js/fixtures/incident.js
@@ -1,0 +1,9 @@
+export function Incident(params = {}) {
+  return {
+    id: '1',
+    title: 'Test Incident',
+    updates: ['First Update', 'Second Update'],
+    url: 'https://status.sentry.io',
+    ...params,
+  };
+}
