+++ tests/js/fixtures/plugins.js
@@ -1,0 +1,16 @@
+import {Plugin} from './plugin';
+
+export function Plugins(params = []) {
+  return [
+    Plugin(),
+    Plugin({
+      enabled: true,
+      id: 'github',
+      name: 'GitHub',
+      slug: 'github',
+      canDisable: false,
+      hasConfiguration: false,
+    }),
+    ...params,
+  ];
+}
