+++ tests/js/fixtures/repository.js
@@ -1,0 +1,10 @@
+export function Repository(params = {}) {
+  return {
+    id: '4',
+    name: 'repo-name',
+    provider: 'github',
+    url: 'https://github.com/example/repo-name',
+    status: 'active',
+    ...params,
+  };
+}
