+++ tests/js/fixtures/team.js
@@ -1,0 +1,9 @@
+export function Team(params) {
+  return {
+    id: '1',
+    slug: 'team-slug',
+    name: 'Team Name',
+    isMember: true,
+    ...params,
+  };
+}
